requirements.txt
streamlit>=1.36
pydantic>=2.0.0
python-dotenv>=1.0.0
